code/bot_license/requirements.txt
cryptography==45.0.6
python-telegram-bot==22.3
